backend/requirements.txt
flask
flask-cors
python-dotenv
gunicorn
requests
openai
httpx
beautifulsoup4
